working version

+++ FileManagerLibrary/Abstractions/IFileWriter.cs
@@ -8,6 +8,6 @@ namespace FileManagerLibrary.Abstractions
     public interface IFileWriter : IDisposable
     {
         void WriteBlock(DataBlock block);
-        void WriteLong(long value);
+        void WriteInt32(int value);
     }
 }
